requirements.txt
crewai>=0.80.0
crewai-tools>=0.14.0
pandas>=2.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
mem0ai>=0.1.0
litellm>=1.0.0
sentence-transformers>=2.2.0
requests>=2.31.0
deepeval>=2.0.0
